#FIX: Separate View and Component
